Apply the review suggestions
