+++ src/index.js
@@ -1,15 +1,14 @@
 import './style.css';
-import add from './modulus/addlist.js';
 import { input, list, out } from './modulus/getelements.js';
-import remove from './modulus/remove.js';
 import display from './modulus/display.js';
-import deleteAll from './modulus/deleteAll';
+import { File } from './modulus/class.js';
 
 document.addEventListener('DOMContentLoaded', () => {
   // Add a TODO
   input.addEventListener('keydown', (e) => {
     if (e.keyCode === 13) {
-      add();
+      const methods = new File();
+      methods.add();
     }
   });
 
@@ -19,12 +18,14 @@ document.addEventListener('DOMContentLoaded', () => {
       const { id } = e.target;
       const l = e.target.parentElement;
       list.removeChild(l);
-      remove(id);
+      const methods = new File();
+      methods.remove(id);
     }
   });
 
   out.addEventListener('click', () => {
-    deleteAll();
+    const methods = new File();
+    methods.deleteAll();
   });
 
   display();
