--- src/modulus/methods.js
@@ -1,10 +1,0 @@
-const methodT = () => {
-  const vl = true;
-  return vl;
-};
-
-const methodF = () => {
-
-};
-
-export default { methodF, methodT };
